[FC] Added SwiftServeInstance to standardize command line interface
Instead of each service parsing the command line itself, a service
can now specify a SwiftServeInstance that includes the root routes
and database specfication.

This instance will handle the command line parsing and therefore
provides a standard interface for running the service as well as
creating and migrating the services database.

It also allows adding aditional command line commands and collecting
extra information necessary for the service to run.